refactor: relocate styled components theme file and refactor code
