+++ src/theme/bootstrap/custom.scss
@@ -1,21 +1,29 @@
-// Override default variables before the import
-
 @import '~bootstrap/scss/functions';
+
+// Theme Color variables override
+$dark: #333333;
+$light: #f2f2f2;
+
 @import '~bootstrap/scss/variables';
+@import '~bootstrap/scss/mixins';
 @import '~bootstrap/scss/utilities';
+
 
 //===========================================
 // Custom Color Map
 //===========================================
 
-$theme-colors: (
-  'gray1': #333333,
+$custom-theme-colors: (
+  'gray1': $dark,
   'gray2': #4f4f4f,
   'gray3': #828282,
   'gray4': #bdbdbd,
   'gray5': #e0e0e0,
-  'gray6': #f2f2f2,
+  'gray6': $light,
 );
+
+$theme-colors: map-merge($custom-theme-colors, $theme-colors);
+
 
 //===========================================
 // Custom Font Size
@@ -26,12 +34,13 @@ $medium-font-size: $font-size-base;
 $small-font-size: $font-size-base * 0.75;
 $overline-font-size: $font-size-base * 0.625;
 
-$font-sizes: (
+$custom-font-sizes: (
   'large': $large-font-size,
   'medium': $medium-font-size,
   'small': $small-font-size,
   'overline': $overline-font-size,
 );
+$font-sizes: map-merge($custom-font-sizes, $font-sizes);
 
 //===========================================
 // Custom line height
@@ -42,21 +51,12 @@ $line-height-medium: 1.5rem;
 $line-height-small: 1.25rem;
 $line-height-overline: 1rem;
 
+
+//===========================================
+// Utilities override
+//===========================================
+
 $utilities: (
-  'line-height': (
-    property: line-height,
-    class: lh,
-    values: (
-      1: 1,
-      sm: $line-height-sm,
-      base: $line-height-base,
-      lg: $line-height-lg,
-      large: $line-height-large,
-      medium: $line-height-medium,
-      small: $line-height-small,
-      overline: $line-height-overline,
-    ),
-  ),
   'color': (
     property: color,
     class: text,
@@ -67,13 +67,34 @@ $utilities: (
           'white': $white,
           'body': $body-color,
           'muted': $text-muted,
-          'black-50': rgba($black, 0.5),
-          'white-50': rgba($white, 0.5),
           'reset': inherit,
         )
       ),
   ),
+  'background-color': (
+    property: background-color,
+    class: bg,
+    values:
+      map-merge(
+        $theme-colors,
+        (
+          'body': $body-bg,
+          'white': $white,
+          'transparent': transparent,
+        )
+      ),
+  ),
+  'line-height': (
+    property: line-height,
+    class: lh,
+    values: (
+      1: 1,
+      large: $line-height-large,
+      medium: $line-height-medium,
+      small: $line-height-small,
+      overline: $line-height-overline,
+    ),
+  ),
 );
 
-// Import Bootstrap and its default variables
 @import '~bootstrap/scss/bootstrap.scss';
