+++ src/theme/styled-components/global.css.ts
@@ -5,49 +5,49 @@ const GlobalStyle = createGlobalStyle`
     box-sizing: border-box;
   }
 
-  html, body {
+  body {
     margin: 0;
     padding: 0;
     font-family: 'Roboto',-apple-system, BlinkMacSystemFont, 'Segoe UI', 'Oxygen', 'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
     sans-serif;
     font-size: 1rem;
     line-height: 1.5rem;
-    background-color: ${(props) => props.theme.color.white};
+    background-color: #fff;
     -webkit-font-smoothing: antialiased;
     -moz-osx-font-smoothing: grayscale;
   }
 
-  h1{
+  h1 {
     font-size: 3.5rem;
     font-weight: 700;
     line-height: 4rem;
   }
 
-  h2{
+  h2 {
     font-size: 3rem;
     font-weight: 700;
     line-height: 3.5rem;
   }
 
-  h3{
+  h3 {
     font-size: 2.5rem;
     font-weight: 700;
     line-height: 3rem;
   }
 
-  h4{
+  h4 {
     font-size: 2rem;
     font-weight: 700;
     line-height: 2.5rem;
   }
 
-  h5{
+  h5 {
     font-size: 1.5rem;
     font-weight: 700;
     line-height: 2rem;
   }
 
-  h6: {
+  h6 {
     font-size: 1rem;
     font-weight: 700;
     line-height: 1.5rem;
