+++ src/theme/styled-components/theme.ts
@@ -1,0 +1,27 @@
+import { DefaultTheme } from 'styled-components';
+
+const theme: DefaultTheme = {
+  color: {
+    gray1: '#333333',
+    gray2: '#4F4F4F',
+    gray3: '#828282',
+    gray4: '#BDBDBD',
+    gray5: '#E0E0E0',
+    gray6: '#F2F2F2',
+  },
+  fontSize: {
+    large: '1.25rem',
+    medium: '1rem',
+    small: '0.75rem',
+    overline: '0.625rem',
+  },
+
+  lineHeight: {
+    large: '1.75rem',
+    medium: '1.5rem',
+    small: '1.25rem',
+    overline: '1rem',
+  },
+};
+
+export default theme;
